Reorganize libCEED integration.
