+++ fem/ceed/integrators/mass/mass.hpp
@@ -12,8 +12,8 @@
 #ifndef MFEM_LIBCEED_MASS_HPP
 #define MFEM_LIBCEED_MASS_HPP
 
-#include "integrator.hpp"
-#include "../fespace.hpp"
+#include "../../interface/integrator.hpp"
+#include "../../../fespace.hpp"
 
 namespace mfem
 {
