+++ fem/ceed/integrators/nlconvection/nlconvection.hpp
@@ -12,8 +12,8 @@
 #ifndef MFEM_LIBCEED_NLCONV_HPP
 #define MFEM_LIBCEED_NLCONV_HPP
 
-#include "integrator.hpp"
-#include "../fespace.hpp"
+#include "../../interface/integrator.hpp"
+#include "../../../fespace.hpp"
 
 namespace mfem
 {
