+++ fem/ceed/interface/interface.hpp
@@ -1,0 +1,35 @@
+// Copyright (c) 2010-2022, Lawrence Livermore National Security, LLC. Produced
+// at the Lawrence Livermore National Laboratory. All Rights reserved. See files
+// LICENSE and NOTICE for details. LLNL-CODE-806117.
+//
+// This file is part of the MFEM library. For more information and source code
+// availability visit https://mfem.org.
+//
+// MFEM is free software; you can redistribute it and/or modify it under the
+// terms of the BSD-3 license. We welcome feedback and contributions, see file
+// CONTRIBUTING.md for details.
+
+#ifndef MFEM_LIBCEED_INTERFACE
+#define MFEM_LIBCEED_INTERFACE
+
+#include "ceed.hpp"
+#include "basis.hpp"
+#include "restriction.hpp"
+#include "operator.hpp"
+#include "integrator.hpp"
+#include "mixed_integrator.hpp"
+#include "coefficient.hpp"
+#include "util.hpp"
+
+namespace mfem
+{
+
+namespace ceed
+{
+
+
+} // namespace ceed
+
+} // namespace mfem
+
+#endif // MFEM_LIBCEED_INTERFACE
