+++ fem/ceed/interface/operator.cpp
@@ -11,9 +11,11 @@
 
 #include "operator.hpp"
 
-#include "../../config/config.hpp"
-#include "../../linalg/vector.hpp"
-#include "../fespace.hpp"
+#include "../../../config/config.hpp"
+#include "../../../linalg/vector.hpp"
+#include "../../fespace.hpp"
+#include "util.hpp"
+#include "ceed.hpp"
 
 namespace mfem
 {
