+++ fem/fem.hpp
@@ -43,7 +43,7 @@
 #include "transfer.hpp"
 #include "fespacehierarchy.hpp"
 #include "multigrid.hpp"
-#include "ceed/algebraic.hpp"
+#include "ceed/solvers/algebraic.hpp"
 #include "lor.hpp"
 
 #ifdef MFEM_USE_MPI
